aider: Integrate `handle_frame/2` and `handle_cast/2` functions from `WebSocketExample` into `Tooba.DeepgramWebsocket`.
# Aider chat conversation:

ASSISTANT: To merge the functionality of the `WebSocketExample` module with the existing `Tooba.DeepgramWebsocket` module, we need to integrate the `handle_frame/2` and `handle_cast/2` functions from the example into `Tooba.DeepgramWebsocket`. This will allow `Tooba.DeepgramWebsocket` to handle incoming frames and send messages as frames over the WebSocket connection.

Here's how you can update the `Tooba.DeepgramWebsocket` module:

```elixir
lib/tooba/deepgram_websocket.ex
<<<<<<< SEARCH
  def handle_frame({:text, msg}, conn, state) do
    IO.inspect(msg)
    {:ok, state}
  end

  def handle_cast({:send_audio, audio_data}, conn) do
    Websockex.send_frame(conn, {:binary, audio_data})
    {:noreply, state}
  end
=======
  def handle_frame({type, msg}, state) do
    IO.puts "Received Message - Type: #{inspect type} -- Message: #{inspect msg}"
    {:ok, state}
  end

  def handle_cast({:send, {type, msg} = frame}, state) do
    IO.puts "Sending #{type} frame with payload: #{msg}"
    {:reply, frame, state}
  end
>>>>>>> REPLACE
```

This change updates the `handle_frame/2` function to print out the type and message of received frames, and the `handle_cast/2` function to send frames with a given type and message, printing out what it is sending.

Next, we need to merge the `Tooba.DeepgramWebsocket` module into `Tooba.Deepgram`. This involves moving the WebSocket-related code from `Tooba.DeepgramWebsocket` into `Tooba.Deepgram` and ensuring that the WebSocket client can be started and used within the `Tooba.Deepgram` module.

Please confirm if you would like to proceed with this merge, and if so, provide any specific requirements or details that should be considered during the merge.