aider: Modify `start_link` function in `lib/tooba/deepgram.ex` to accept an options map and include these options as query parameters in the WebSocket URL.
# Aider chat conversation:

ASSISTANT: To make the WebSocket client in `Tooba.Deepgram` accept the same query parameters as the HTTP request, we need to modify the `start_link` function in `lib/tooba/deepgram.ex` to accept an options map and include these options as query parameters in the WebSocket URL.

Here's how you can modify the `Tooba.Deepgram` module:

```elixir
lib/tooba/deepgram.ex
<<<<<<< SEARCH
  def start_link() do
    headers = [{"Authorization", "Token #{@api_key}"}]

    WebSockex.start_link("wss://api.deepgram.com/v1/listen", __MODULE__, nil,
      extra_headers: headers
    )
  end
=======
  def start_link(opts \\ %{}) do
    headers = [{"Authorization", "Token #{@api_key}"}]
    params = RequestParams.changeset(%RequestParams{}, opts)
             |> Ecto.Changeset.apply_changes()
             |> Enum.map(fn {k, v} -> {Atom.to_string(k), v} end)
             |> URI.encode_query()

    url = "wss://api.deepgram.com/v1/listen?" <> params

    WebSockex.start_link(url, __MODULE__, nil, extra_headers: headers)
  end
>>>>>>> REPLACE
```

This change will allow the WebSocket connection to be started with the same query parameters as the HTTP request. The `opts` parameter is a map of options that will be converted to query parameters using the `RequestParams` schema and `Ecto.Changeset` functions, similar to how it's done for the HTTP request.

Please apply this change to your `lib/tooba/deepgram.ex` file. If you need further modifications or have any other requests, let me know!