aider: Pass `deepgram_opts` as a parameter to `Tooba.DeepgramSink` and handle the initialization with the given options.
# Aider chat conversation:

ASSISTANT: To make the `deepgram_opts` a parameter of the `Tooba.DeepgramSink` sink, we need to modify the `Tooba.Record` module to pass the options when creating the sink element in the pipeline. We will also need to adjust the `Tooba.DeepgramSink` module to accept and use these options.

Let's start by modifying the `Tooba.Record` module to pass the `deepgram_opts` as a parameter to the `Tooba.DeepgramSink` element.

```elixir
lib/tooba/record.ex
<<<<<<< SEARCH
      |> child(Tooba.DeepgramSink, deepgram_opts)
=======
      |> child(Tooba.DeepgramSink, %{deepgram_opts: deepgram_opts})
>>>>>>> REPLACE
```

Next, we need to update the `Tooba.DeepgramSink` module to handle the initialization with the given options.

```elixir
lib/tooba/deepgram_sink.ex
<<<<<<< SEARCH
  def handle_init(opts) do
    # Start the WebSocket client and keep the pid to send messages later
    {:ok, ws_pid} = Deepgram.start_link(opts)
=======
  def handle_init(%{deepgram_opts: deepgram_opts}) do
    # Start the WebSocket client and keep the pid to send messages later
    {:ok, ws_pid} = Deepgram.start_link(deepgram_opts)
>>>>>>> REPLACE
    {:ok, %{ws_pid: ws_pid}}
  end
```

With these changes, the `deepgram_opts` will be passed as a parameter to the `Tooba.DeepgramSink` element, and the `handle_init` function will extract the options and use them to start the WebSocket client.