aider: Added `def_input_pad` macro to define an input pad for the `Tooba.DeepgramSink` module.
# Aider chat conversation:

ASSISTANT: To insert the `def_input_pad` into the `Tooba.DeepgramSink` module, we will need to modify the `lib/tooba/deepgram_sink.ex` file. The `def_input_pad` macro is used to define an input pad for a Membrane element, which is necessary for a sink element to receive data.

Here's the change that needs to be made:

```elixir
lib/tooba/deepgram_sink.ex
<<<<<<< SEARCH
defmodule Tooba.DeepgramSink do
  use Membrane.Sink
=======
defmodule Tooba.DeepgramSink do
  use Membrane.Sink

  def_input_pad :input,
    demand_unit: :buffers,
    caps: :any,
    availability: :on_request,
    mode: :pull
>>>>>>> REPLACE
```

This code defines an input pad named `:input` with the specified options. The `demand_unit` is set to `:buffers`, which means the element will demand and process data in buffer units. The `caps` is set to `:any`, indicating that the sink can handle any type of data. The `availability` is set to `:on_request`, which means the pad will be available when requested by the upstream element. The `mode` is set to `:pull`, which means the sink will pull data from the upstream element.

Please confirm if you would like to proceed with this change, or if there are any additional details or modifications you would like to include.